Move battery sync specific code out of WatchMessageReceiver.kt Functions now live in Utils.kt
